requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
python-rtmidi>=1.5.0
numpy>=1.24.0
